[feature/homepage-layout]: Adjusting bubbles to stand alone component
