+++ src/views/Home/Home.test.js
@@ -1,0 +1,14 @@
+import { shallowMount } from "@vue/test-utils";
+import Home from "@/views/Home/Home.vue";
+
+describe.skip("Home Page :: unit tests", () => {
+  const wrapper = shallowMount(Home, {
+    props: {
+      // linkCollection: mockLinkCollection,
+    },
+  });
+
+  it.skip("Component renders", () => {
+    expect(wrapper.exists()).toBe(true);
+  });
+});
