+++ src/views/Home/Home.vue
@@ -24,35 +24,28 @@
       </div>
       <SiteNav customFontClass="bold" />
     </div>
-    <div
-      class="bubble"
-      v-for="n in TOTAL_BUBBLES"
-      :key="n"
-      v-memo="[TOTAL_BUBBLES]"
-    ></div>
+    <BackgroundBubbles />
   </div>
 </template>
 
 <script>
 import SiteNav from "@/components/SiteNav/SiteNav.vue";
+import BackgroundBubbles from "@components/BackgroundBubbles/BackgroundBubbles.vue";
 export default {
   name: "Home",
   components: {
     SiteNav,
+    BackgroundBubbles,
   },
-  props: {
-    msg: String,
-  },
+  props: {},
   data() {
     return {
       modalVisibility: false,
-      TOTAL_BUBBLES: 59,
     };
   },
   methods: {
     handleOnClick(event) {
       if (event) {
-        console.log("Home::handleOnClickEvent::event::", event);
         this.toggleModalVisibility();
       }
     },
@@ -64,7 +57,6 @@ export default {
 </script>
 <style lang="scss" scoped>
 //temp
-
 .home-wrap {
   height: 100%;
   margin: 0;
@@ -81,6 +73,7 @@ export default {
 ::v-deep {
   .nav-wrap {
     flex: 2;
+
     ul {
       list-style-type: none;
       padding-left: 0;
@@ -90,21 +83,18 @@ export default {
 }
 
 .flex-container {
-  position: relative;
-  z-index: 2; // Foreground content layer
+  position: absolute;
+  top: 25%;
+  z-index: 2;
   display: flex;
   justify-content: center;
   align-items: center;
-  // height: 100vh;
   width: 100%;
   padding: 3em;
 
   .copy-wrap {
     flex-flow: row;
-    // padding: 20px;
-    // background-color: #f0f0f0;
-    // border-radius: 8px;
-    // box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    display: flex;
 
     header {
       display: flex;
@@ -115,8 +105,6 @@ export default {
       text-align: right;
       padding-left: 5em;
       padding: 0 3em;
-
-      // text-align: right;
 
       p {
         color: #ffff;
@@ -137,97 +125,28 @@ export default {
       font-style: normal;
       line-height: 0.9em;
     }
+
     b {
       font-weight: 700;
     }
   }
 }
 
-.copy-wrap {
-  display: flex;
-}
+@media only screen and (max-width: 600px) {
+  .flex-container {
+    flex-flow: column;
+    justify-content: flex-start;
+    align-items: flex-start;
 
-$bubble-count: 50;
-$sway-type: "sway-left-to-right", "sway-right-to-left";
+    .copy-wrap {
+      flex-flow: column;
+      width: 100%;
 
-@function random_range($min, $max) {
-  $rand: random();
-  $random_range: $min + floor($rand * (($max - $min) + 1));
-  @return $random_range;
-}
-
-@function sample($list) {
-  @return nth($list, random(length($list)));
-}
-
-.bubbles {
-  position: relative;
-  width: 100%;
-  height: 100vh;
-  overflow: hidden;
-  mouse-interactions: none;
-}
-
-.bubble {
-  position: absolute;
-  left: var(--bubble-left-offset);
-  bottom: -75%;
-  display: block;
-  width: var(--bubble-radius);
-  height: var(--bubble-radius);
-  border-radius: 50%;
-  animation: float-up var(--bubble-float-duration) var(--bubble-float-delay)
-    ease-in infinite;
-
-  &::before {
-    position: absolute;
-    content: "";
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: hsla(240, 72%, 52%, 0.3);
-    border-radius: inherit;
-    animation: var(--bubble-sway-type) var(--bubble-sway-duration)
-      var(--bubble-sway-delay) ease-in-out alternate infinite;
-  }
-
-  @for $i from 0 through $bubble-count {
-    &:nth-child(#{$i}) {
-      --bubble-left-offset: #{random_range(0vw, 100vw)};
-      --bubble-radius: #{random_range(1vw, 10vw)};
-      --bubble-float-duration: #{random_range(6s, 12s)};
-      --bubble-sway-duration: #{random_range(4s, 6s)};
-      --bubble-float-delay: #{random_range(0s, 4s)};
-      --bubble-sway-delay: #{random_range(0s, 4s)};
-      --bubble-sway-type: #{sample($sway-type)};
+      header {
+        text-align: inherit;
+        padding: 0;
+      }
     }
-  }
-}
-
-@keyframes float-up {
-  to {
-    transform: translateY(-175vh);
-  }
-}
-
-@keyframes sway-left-to-right {
-  from {
-    transform: translateX(-100%);
-  }
-
-  to {
-    transform: translateX(100%);
-  }
-}
-
-@keyframes sway-right-to-left {
-  from {
-    transform: translateX(100%);
-  }
-
-  to {
-    transform: translateX(-100%);
   }
 }
 </style>
